# refactor(sdk): tweak StepUpload finish condition so it exits faster