Description
-----------

- I think the new code is more understandable!

- It's a functional no-op, for non-obvious reasons:

    - Technical change 1: ignoring non-`EventJobDone` messages after the RequestFinish. This is fine because [RequestFinish is guaranteed to be the last message put into the queue](https://github.com/wandb/wandb/blob/d244d07f4eb2bf9af0d247c78f04320dc2129955/wandb/filesync/step_checksum.py#L139)
    - Technical change 2: if `not self._running_jobs` on the first evaluation of the while-condition, the loop will get skipped entirely. This is fine because `self._running_jobs` is only ever extended _synchronously, in `_handle_event`_ -- which means that, if the earlier `_handle_event` calls didn't append to `self._running_jobs`, there's no reason to wait 0.2s before seeing that it's empty and exiting.

- The tests run much faster, because there used to be a 0.2s sleep every time a StepUpload shut down.

    Before:
    
    ```
    $ pytest --durations=0 tests/unit_tests/test_step_upload.py
    <...snip...>
    =================================================== slowest durations ====================================================
    0.32s call     tests/unit_tests/test_step_upload.py::test_is_alive_until_last_job_finishes
    0.31s call     tests/unit_tests/test_step_upload.py::test_enforces_max_jobs
    0.31s call     tests/unit_tests/test_step_upload.py::TestArtifactCommit::test_no_commit_until_uploads_done
    0.31s call     tests/unit_tests/test_step_upload.py::TestUpload::test_reuploads_if_event_during_upload
    0.31s call     tests/unit_tests/test_step_upload.py::TestFinish::test_no_finish_until_jobs_done
    0.30s setup    tests/unit_tests/test_step_upload.py::TestFinish::test_finishes[commands0]
    0.23s call     tests/unit_tests/test_step_upload.py::TestArtifactCommit::test_no_commit_if_upload_fails
    0.22s call     tests/unit_tests/test_step_upload.py::TestFinish::test_finishes[commands1]
    0.21s call     tests/unit_tests/test_step_upload.py::TestUpload::test_upload
    0.21s call     tests/unit_tests/test_step_upload.py::TestUpload::test_deletes_after_upload_iff_copied[True]
    0.21s call     tests/unit_tests/test_step_upload.py::TestFinish::test_finishes[commands2]
    0.21s call     tests/unit_tests/test_step_upload.py::TestFinish::test_finishes[commands0]
    0.21s call     tests/unit_tests/test_step_upload.py::TestArtifactCommit::test_calls_callbacks
    0.21s call     tests/unit_tests/test_step_upload.py::TestArtifactCommit::test_commits_iff_finalize[False]
    0.21s call     tests/unit_tests/test_step_upload.py::TestArtifactCommit::test_commits_iff_finalize[True]
    0.20s call     tests/unit_tests/test_step_upload.py::TestUpload::test_deletes_after_upload_iff_copied[False]
    0.20s call     tests/unit_tests/test_step_upload.py::TestFinish::test_finishes[commands3]
    0.20s call     tests/unit_tests/test_step_upload.py::TestFinish::test_finishes[commands4]
    0.01s teardown tests/unit_tests/test_step_upload.py::test_is_alive_until_last_job_finishes
    0.00s setup    tests/unit_tests/test_step_upload.py::test_enforces_max_jobs
    0.00s setup    tests/unit_tests/test_step_upload.py::TestUpload::test_deletes_after_upload_iff_copied[True]
    <...snip...>
    ```
    
    After:
    
    ```
    $ pytest --durations=0 tests/unit_tests/test_step_upload.py
    <...snip...>
    =================================================== slowest durations ====================================================
    0.35s setup    tests/unit_tests/test_step_upload.py::TestFinish::test_finishes[commands0]
    0.11s call     tests/unit_tests/test_step_upload.py::TestUpload::test_reuploads_if_event_during_upload
    0.11s call     tests/unit_tests/test_step_upload.py::TestArtifactCommit::test_no_commit_until_uploads_done
    0.11s call     tests/unit_tests/test_step_upload.py::test_enforces_max_jobs
    0.10s call     tests/unit_tests/test_step_upload.py::test_is_alive_until_last_job_finishes
    0.10s call     tests/unit_tests/test_step_upload.py::TestFinish::test_no_finish_until_jobs_done
    0.01s call     tests/unit_tests/test_step_upload.py::TestArtifactCommit::test_no_commit_if_upload_fails
    0.01s call     tests/unit_tests/test_step_upload.py::TestFinish::test_finishes[commands1]
    <...snip...>
    ```



Testing
-------
Our existing unit tests exercise StepUpload's finish-logic pretty heavily, including e.g. asserting that it doesn't exit before all of its jobs have finished.